+++ app/src/main/java/org/breezyweather/common/basic/models/options/appearance/DailyTrendDisplay.kt
@@ -29,8 +29,9 @@ enum class DailyTrendDisplay(
     TAG_TEMPERATURE("temperature", R.string.temperature), 
     TAG_AIR_QUALITY("air_quality", R.string.air_quality),
     TAG_WIND("wind", R.string.wind), 
-    TAG_UV_INDEX("uv_index", R.string.uv_index), 
-    TAG_PRECIPITATION("precipitation", R.string.precipitation);
+    TAG_UV_INDEX("uv_index", R.string.uv_index),
+    TAG_PRECIPITATION("precipitation", R.string.precipitation),
+    TAG_FEELS_LIKE("feels_like", R.string.temperature_feels_like);
 
     companion object {
 
@@ -48,6 +49,7 @@ enum class DailyTrendDisplay(
                     "wind" -> list.add(TAG_WIND)
                     "uv_index" -> list.add(TAG_UV_INDEX)
                     "precipitation" -> list.add(TAG_PRECIPITATION)
+                    "feels_like" -> list.add(TAG_FEELS_LIKE)
                 }
             }
             list