+++ app/src/main/java/org/breezyweather/main/adapters/trend/hourly/HourlyHumidityAdapter.kt
@@ -1,0 +1,189 @@
+/**
+ * This file is part of Breezy Weather.
+ *
+ * Breezy Weather is free software: you can redistribute it and/or modify it
+ * under the terms of the GNU Lesser General Public License as published by the
+ * Free Software Foundation, version 3 of the License.
+ *
+ * Breezy Weather is distributed in the hope that it will be useful, but
+ * WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY
+ * or FITNESS FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public
+ * License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with Breezy Weather. If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package org.breezyweather.main.adapters.trend.hourly
+
+import android.content.Context
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.annotation.Size
+import org.breezyweather.R
+import org.breezyweather.common.basic.GeoActivity
+import org.breezyweather.common.basic.models.Location
+import org.breezyweather.common.basic.models.options.unit.ProbabilityUnit
+import org.breezyweather.common.basic.models.options.unit.TemperatureUnit
+import org.breezyweather.common.basic.models.weather.Temperature
+import org.breezyweather.common.ui.widgets.trend.TrendRecyclerView
+import org.breezyweather.common.ui.widgets.trend.chart.PolylineAndHistogramView
+import org.breezyweather.main.utils.MainThemeColorProvider
+import org.breezyweather.theme.ThemeManager
+import org.breezyweather.theme.resource.ResourceHelper
+import org.breezyweather.theme.resource.providers.ResourceProvider
+import org.breezyweather.theme.weatherView.WeatherViewController
+import kotlin.math.max
+
+/**
+ * Hourly humidity / dew point adapter.
+ */
+class HourlyHumidityAdapter(
+    activity: GeoActivity,
+    location: Location,
+    provider: ResourceProvider,
+    unit: TemperatureUnit
+) : AbsHourlyTrendAdapter(activity, location) {
+    private val mResourceProvider: ResourceProvider = provider
+    private val mDewPointUnit: TemperatureUnit = unit
+    private val mDewPoints: Array<Float?>
+    private var mHighestDewPoint: Float? = null
+    private var mLowestDewPoint: Float? = null
+
+    inner class ViewHolder(itemView: View) : AbsHourlyTrendAdapter.ViewHolder(itemView) {
+        private val mPolylineAndHistogramView = PolylineAndHistogramView(itemView.context)
+
+        init {
+            hourlyItem.chartItemView = mPolylineAndHistogramView
+        }
+
+        fun onBindView(activity: GeoActivity, location: Location, position: Int) {
+            val talkBackBuilder = StringBuilder(activity.getString(R.string.tag_humidity_dew_point))
+            super.onBindView(activity, location, talkBackBuilder, position)
+            val weather = location.weather!!
+            val hourly = weather.nextHourlyForecast[position]
+            if (hourly.dewPoint != null) {
+                talkBackBuilder.append(", ").append(Temperature.getTemperature(activity, hourly.dewPoint, mDewPointUnit))
+            }
+            hourlyItem.setIconDrawable(
+                hourly.weatherCode?.let {
+                    ResourceHelper.getWeatherIcon(mResourceProvider, it, hourly.isDaylight)
+                },
+                missingIconVisibility = View.INVISIBLE
+            )
+            val relativeHumidity = hourly.relativeHumidity
+            mPolylineAndHistogramView.setData(
+                buildDewPointArrayForItem(mDewPoints, position),
+                null,
+                Temperature.getShortTemperature(activity, hourly.dewPoint, mDewPointUnit),
+                null,
+                mHighestDewPoint,
+                mLowestDewPoint,
+                relativeHumidity,
+                if (relativeHumidity != null) ProbabilityUnit.PERCENT.getValueText(activity, relativeHumidity.toInt()) else null,
+                100f,
+                0f
+            )
+            val themeColors = ThemeManager
+                .getInstance(itemView.context)
+                .weatherThemeDelegate
+                .getThemeColors(
+                    itemView.context,
+                    WeatherViewController.getWeatherKind(location.weather),
+                    location.isDaylight
+                )
+            val lightTheme = MainThemeColorProvider.isLightTheme(itemView.context, location)
+            mPolylineAndHistogramView.setLineColors(
+                themeColors[if (lightTheme) 1 else 2],
+                themeColors[2],
+                MainThemeColorProvider.getColor(location, com.google.android.material.R.attr.colorOutline)
+            )
+            mPolylineAndHistogramView.setShadowColors(
+                themeColors[if (lightTheme) 1 else 2],
+                themeColors[2],
+                lightTheme
+            )
+            mPolylineAndHistogramView.setTextColors(
+                MainThemeColorProvider.getColor(location, R.attr.colorTitleText),
+                MainThemeColorProvider.getColor(location, R.attr.colorBodyText),
+                MainThemeColorProvider.getColor(location, R.attr.colorPrecipitationProbability)
+            )
+            mPolylineAndHistogramView.setHistogramAlpha(if (lightTheme) 0.2f else 0.5f)
+            hourlyItem.contentDescription = talkBackBuilder.toString()
+        }
+
+        @Size(3)
+        private fun buildDewPointArrayForItem(temps: Array<Float?>, adapterPosition: Int): Array<Float?> {
+            val a = arrayOfNulls<Float>(3)
+            a[1] = temps[2 * adapterPosition]
+            if (2 * adapterPosition - 1 < 0) {
+                a[0] = null
+            } else {
+                a[0] = temps[2 * adapterPosition - 1]
+            }
+            if (2 * adapterPosition + 1 >= temps.size) {
+                a[2] = null
+            } else {
+                a[2] = temps[2 * adapterPosition + 1]
+            }
+            return a
+        }
+    }
+
+    init {
+        val weather = location.weather!!
+        mDewPoints = arrayOfNulls(max(0, weather.nextHourlyForecast.size * 2 - 1))
+        run {
+            var i = 0
+            while (i < mDewPoints.size) {
+                mDewPoints[i] = weather.nextHourlyForecast.getOrNull(i / 2)?.dewPoint
+                i += 2
+            }
+        }
+        run {
+            var i = 1
+            while (i < mDewPoints.size) {
+                if (mDewPoints[i - 1] != null && mDewPoints[i + 1] != null) {
+                    mDewPoints[i] = (mDewPoints[i - 1]!! + mDewPoints[i + 1]!!) * 0.5f
+                } else {
+                    mDewPoints[i] = null
+                }
+                i += 2
+            }
+        }
+        weather.nextHourlyForecast
+            .forEach { hourly ->
+                hourly.dewPoint?.let {
+                    if (mHighestDewPoint == null || it > mHighestDewPoint!!) {
+                        mHighestDewPoint = it
+                    }
+                    if (mLowestDewPoint == null || it < mLowestDewPoint!!) {
+                        mLowestDewPoint = it
+                    }
+                }
+            }
+    }
+
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
+        val view = LayoutInflater.from(parent.context)
+            .inflate(R.layout.item_trend_hourly, parent, false)
+        return ViewHolder(view)
+    }
+
+    override fun onBindViewHolder(holder: AbsHourlyTrendAdapter.ViewHolder, position: Int) {
+        (holder as ViewHolder).onBindView(activity, location, position)
+    }
+
+    override fun getItemCount() = location.weather!!.nextHourlyForecast.size
+
+    override fun isValid(location: Location): Boolean {
+        return mHighestDewPoint != null && mLowestDewPoint != null
+    }
+
+    override fun getDisplayName(context: Context) = context.getString(R.string.tag_humidity_dew_point)
+
+    override fun bindBackgroundForHost(host: TrendRecyclerView) {
+        host.setData(null, 0f, 0f)
+    }
+}